pyproject.toml
[tool.poetry]
name = "mwdstdb"
version = "0.5.0"
license = "AGPL-3.0-or-later"
description = "REST storage and API backend"
authors = ["MWD STD Inc. <info@mwdstd.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "0.70.0"
celery = "4.4.7"
celery-pool-asyncio = "0.2.0"
motor = "2.1.0"
uvicorn = "0.11.3"
Pint = "^0.19.2"
jsons = "1.6.3"
httpx = "0.14.1"
python-jose = "3.2.0"
numpy = "^1.23.1"
python-multipart = "^0.0.5"
sse-starlette = "^0.10.3"
redis = "^4.3.4"
pymongo = {version = "3.10.1", extras = ["srv"]}
gunicorn = {version = "^19.9.0", optional = true}

[tool.poetry.dev-dependencies]

[tool.poetry.extras]
gunicorn = ["gunicorn"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
